Refactor deleteVm command into own folder
